Configured graph encoder to include gru messenger
