+++ tests/domain/test_graph_encoder.py
@@ -2,9 +2,8 @@ from unittest import TestCase
 
 import numpy as np
 
-from src.domain.graph import Graph
 from src.domain.graph_encoder import GraphEncoder
-from src.domain.messenger_gru import MessengerGRU
+from src.domain.graph import Graph
 from tests.fixtures.matrices_and_vectors import BASE_GRAPH, BASE_GRAPH_NODE_FEATURES, \
     BASE_U_MATRIX, BASE_GRAPH_EDGE_FEATURES, BASE_W_MATRIX, MULTIPLICATION_FACTOR, BASE_B_VECTOR
 
@@ -12,23 +11,22 @@ from tests.fixtures.matrices_and_vectors import BASE_GRAPH, BASE_GRAPH_NODE_FEAT
 class TestGraphEncoder(TestCase):
 
     def setUp(self) -> None:
-        self.gru_messenger = MessengerGRU()
-        self.gru_messenger.w_tree_update_gate_features = MULTIPLICATION_FACTOR * BASE_W_MATRIX
-        self.gru_messenger.w_tree_forget_gate_features = MULTIPLICATION_FACTOR * BASE_W_MATRIX
-        self.gru_messenger.w_tree_current_memory_message_features = MULTIPLICATION_FACTOR * BASE_W_MATRIX
-        self.gru_messenger.u_tree_update_gate = MULTIPLICATION_FACTOR * BASE_W_MATRIX
-        self.gru_messenger.u_tree_forget_gate = MULTIPLICATION_FACTOR * BASE_W_MATRIX
-        self.gru_messenger.u_tree_current_memory_message = MULTIPLICATION_FACTOR * BASE_W_MATRIX
-        self.gru_messenger.b_tree_update_gate = MULTIPLICATION_FACTOR * BASE_B_VECTOR
-        self.gru_messenger.b_tree_forget_gate = MULTIPLICATION_FACTOR * BASE_B_VECTOR
-        self.gru_messenger.b_tree_current_memory_message = MULTIPLICATION_FACTOR * BASE_B_VECTOR
-        self.graph_encoder = GraphEncoder(self.gru_messenger)
+        self.graph_encoder = GraphEncoder()
+        self.graph_encoder.w_gru_update_gate_features = MULTIPLICATION_FACTOR * BASE_W_MATRIX
+        self.graph_encoder.w_gru_forget_gate_features = MULTIPLICATION_FACTOR * BASE_W_MATRIX
+        self.graph_encoder.w_gru_current_memory_message_features = MULTIPLICATION_FACTOR * BASE_W_MATRIX
+        self.graph_encoder.u_gru_update_gate = MULTIPLICATION_FACTOR * BASE_W_MATRIX
+        self.graph_encoder.u_gru_forget_gate = MULTIPLICATION_FACTOR * BASE_W_MATRIX
+        self.graph_encoder.u_gru_current_memory_message = MULTIPLICATION_FACTOR * BASE_W_MATRIX
+        self.graph_encoder.b_gru_update_gate = MULTIPLICATION_FACTOR * BASE_B_VECTOR
+        self.graph_encoder.b_gru_forget_gate = MULTIPLICATION_FACTOR * BASE_B_VECTOR
+        self.graph_encoder.b_gru_current_memory_message = MULTIPLICATION_FACTOR * BASE_B_VECTOR
         self.graph_encoder.u_graph_node_features = 0.1 * BASE_U_MATRIX
         self.graph_encoder.u_graph_neighbor_messages = 0.1 * BASE_U_MATRIX
 
     def test_encode_graph_returns_the_expected_encoding_for_a_node_after_one_time_step(self):
         # Given
-        self.gru_messenger.time_steps = 1
+        self.graph_encoder.time_steps = 1
         node = 0
         node_encoding_expected = np.array([0.3909883, 0.3909883])
         graph = Graph(BASE_GRAPH,
@@ -43,7 +41,7 @@ class TestGraphEncoder(TestCase):
 
     def test_encode_graph_returns_the_expected_shape(self):
         # Given
-        self.gru_messenger.time_steps = 1
+        self.graph_encoder.time_steps = 1
         encoded_graph_shape_expected = BASE_GRAPH_NODE_FEATURES.shape
         graph = Graph(BASE_GRAPH,
                       BASE_GRAPH_NODE_FEATURES,
